Add React Component To Display Message
Add a React Component to display a message that it was successful.

A React Component is a modular in React where a single part of the
application is created it renders to the screen. The render uses uses a
XMLish type of tags called JSX. Which simple precompilers than
translates the syntactic sugar to plain JavaScript. So in actually it
would look like this with use of hashes and function createElement in
order for JavaScript to use it.

var CommentBox = React.createClass({displayName: 'CommentBox',
  render: function() {
    return (
      React.createElement('div', {className: "commentBox"},
        "Hello, world! I am a CommentBox."
      )
    );
  }
});
React.render(
  React.createElement(CommentBox, null),
  document.getElementById('content')
);

You can also add in additional other React components into render
allowing for nesting of components and isolated components to keep code
easy to maintain.