Add React Component Comment
Add the React Component Comment which will display the comments of the
user. Also, this component uses Props, they are like public valuables
where the developer can access them and set the valuables for them.

In order to do this add them to the React Component tag as a attribute
with the name of the prop and the value you would like to set. In this
cast <Comment author="Jason">This comment</comment>. Furthermore, by
using this.props.children you can access the text or other tags between
the opening and closing tags.