Add Raw HTML Display For Comments
Because React protect its developers from XSS attacks by default it
turns any HTML into harmless strings. This can be override and as long
you know that the source of the HTML is secure. By adding
dangerouslySetInnerHTML attribute to a HTML tag you can set the inter
HTML to render as HTML in the browser.