Change Props.data To State.data In CommentBox
Change Prop.data to State.data in CommentBox since a State is like a
private valuable for a component it can only be access from that
component. In order to set a state you will need to add getInitialState
to that component and declare that state valuables. In order to set a
state outside of a getInitalSate it is this.setState({stateName: value})
this will course the component to re-render in order to keep it up to
data with the latest data from the state. To access those state it is
similar to props this.state.stateName.