Add Submit Event Handler To Comment Form
Add in the submit handler to the Comment Form to submit a new comment to
the server. By using the prefix _on to tell that this is an event
handler which should only be used inside this component. Also adding
preventDefault function to prevent the default action of submit from
occurring. This handler uses this.refs Reacts way of find tags in the
render and findDOMNode in order to get the properties and functions for
those tags.