Add Callback Function To CommentBox
Add in a callback function to send back to the CommentBox the author and
text in order for the CommentBox to send it to the server. Because
CommentBox controls displaying the list of comments and it should
display the newest comment once it has been complete. This way the
CommentForm only is responsible for get and verifying the data from the
user. CommentBox is responsible for displaying the comments.